requirements.txt
django==2.2.17
djangorestframework==3.12.2
Pillow==8.0.1
sorl-thumbnail==12.6.3

# https://pypi.org/project/django-stubs/ https://habr.com/ru/post/465007/